requirements.txt
fastmcp>=2.13.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
openai>=1.6.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
schedule>=1.2.0
pydantic>=2.5.0
